Change spacing using prettier so lint doesn't hate my tabs.
