--- src/api/api.ts
@@ -1,5 +1,0 @@
-import twillio from './twillioIntegration';
-
-export default {
-    twillio
-}
